+++ src/components/galleryPage/modalButton/ModalButton.jsx
@@ -1,4 +1,4 @@
-import { StyledModalButton } from '../../../styled/galleryPage/modalButton/StyledModalButton';
+import { StyledModalButton } from '../../../styled/galleryPage/StyledModalButton';
 
 export const ModalButton = () => (
     <StyledModalButton>Оставить комментарий</StyledModalButton>
