+++ src/styled/galleryPage/StyledModalButton.jsx
@@ -1,6 +1,6 @@
 import styled from 'styled-components';
 
-import { variables } from '../../variables';
+import { variables } from '../variables';
 
 export const StyledModalButton = styled.button`
     background-color: ${variables.color.actionBg};
