--- src/styled/galleryPage/galleryImage/StyledGalleryImage.jsx
@@ -1,8 +1,0 @@
-import styled from 'styled-components';
-
-export const StyledGalleryImage = styled.img.attrs((props) => ({
-    src: props.src || null
-}))`
-    width: 100%;
-    cursor: pointer;
-`;
